Renamed media container in typescript
